sbom/requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
requests>=2.25.0
packaging>=21.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
toml>=0.10.0 